Refactor QuotedStringTokenizer
